Fix bugs and dead weight found reading the graphics library
Net -76 lines in libraries/graphics. All 43 goldens stay byte-identical and the
380 tests pass, so nothing here changes a pixel.

Bugs:

- `TextureAtlas::new` summed each surface's *height* into the atlas width as
  well as its width. Packing was still correct, so this was invisible - the
  texture was just far wider than anything drawn into it, several megabytes of
  transparent pixels on the GPU for the block and wall atlases.
- `AnimationTimer` held absolute milliseconds since construction in an `i32`.
  After 24.8 days of uptime that overflows, the elapsed comparison goes
  permanently false, and every animation in the game stops. `Button` and
  `Toggle` kept the same counter in a `u32`, which does it at 49.7 days. Both
  are 64 bit now.
- `Surface`'s mutable iterator handed out borrows through a raw pointer because
  the compiler could not see they were disjoint. Iterating the backing slice
  instead is safe, shorter and faster - `iter_mut` already knows. That was the
  library's only `unsafe`.

Optimizations:

- `TextInput` measured the text either side of the cursor by rasterising it into
  a throwaway `Surface`, twice, on every frame the input was selected.
  `get_text_size` is the measuring half of that same function and returns the
  same number without drawing a glyph.
- `RenderRect` recorded a blur and a nine-quad shadow even when both were
  switched off, which is the default. Neither drew anything - the backend
  discards a sub-pixel blur radius and a zero intensity shadow is tinted to zero
  alpha - so this is 18 fewer commands per in-game frame, 345 down to 327.
- `ShadowContext::new` filled its 700x700 surface and then faded it in a second
  pass over the same half million pixels. One pass does both.
- `surface_bytes` walked the bounds-checked iterator to build an upload buffer
  when it could walk the pixel slice.

Removed: `Container::update_subelements`, which nothing ever called and which
was hiding a dead `buttons_container` in two menus; `MeshHandle::get_vertex_count`;
and four hand-rolled colour lerps that `interpolate_colors` already does.

Also corrected the doc comments that still described a `gl_backend::GlBackend`,
a `gpu_garbage` module and SDL events, none of which exist.

+++ libraries/graphics/animation_timer.rs
@@ -1,18 +1,22 @@
-/// `AnimationTimer` has one function: `is_frame_ready`.
+/// `AnimationTimer` has one function: `frame_ready`.
 ///
 /// It accumulates the time elapsed, for example: if you wait 1000ms and duration is 100ms,
-/// then first 10 times `is_frame_ready` will return true and 11th time it will return false.
+/// then first 10 times `frame_ready` will return true and 11th time it will return false.
+///
+/// The counters are 64 bit because they are absolute milliseconds since the timer was made,
+/// not a delta. As `i32` they overflowed after 24.8 days of uptime, at which point the
+/// elapsed time compared as negative and every animation in the game stopped for good.
 #[derive(Debug)]
 pub struct AnimationTimer {
-    per_frame: i32,
+    per_frame: i64,
     start_time: std::time::Instant,
-    ms_passed: i32,
+    ms_passed: i64,
 }
 
 impl AnimationTimer {
     /// Creates a new `AnimationTimer` with the given duration.
     #[must_use]
-    pub fn new(per_frame: i32) -> Self {
+    pub fn new(per_frame: i64) -> Self {
         Self {
             per_frame,
             start_time: std::time::Instant::now(),
@@ -27,13 +31,13 @@ impl AnimationTimer {
     /// reach that widget.
     #[cfg(feature = "render-tests")]
     pub const fn freeze(&mut self) {
-        self.ms_passed = i32::MAX;
+        self.ms_passed = i64::MAX;
     }
 
     /// Returns true if the frame is ready to be rendered.
     #[must_use]
     pub fn frame_ready(&mut self) -> bool {
-        let elapsed = self.start_time.elapsed().as_millis() as i32;
+        let elapsed = self.start_time.elapsed().as_millis() as i64;
         if elapsed > self.ms_passed {
             self.ms_passed += self.per_frame;
             true
